+++ pkg/engine/datasource/graphql_datasource/graphql_http_datasource/graphql_subscription_client_test.go
@@ -1,4 +1,4 @@
-package graphql_datasource
+package graphql_http_datasource
 
 import (
 	"context"
@@ -13,6 +13,7 @@ import (
 
 	"github.com/buger/jsonparser"
 	ll "github.com/jensneuse/abstractlogger"
+	"github.com/jensneuse/graphql-go-tools/pkg/engine/datasource/graphql_datasource"
 	"github.com/stretchr/testify/assert"
 	"go.uber.org/atomic"
 	"go.uber.org/zap"
@@ -68,9 +69,9 @@ func TestWebsocketSubscriptionClient(t *testing.T) {
 		WithLogger(logger()),
 	)
 	next := make(chan []byte)
-	err := client.Subscribe(ctx, GraphQLSubscriptionOptions{
+	err := client.Subscribe(ctx, graphql_datasource.GraphQLSubscriptionOptions{
 		URL: strings.Replace(server.URL, "http", "ws", -1),
-		Body: GraphQLBody{
+		Body: graphql_datasource.GraphQLBody{
 			Query: `subscription {messageAdded(roomName: "room"){text}}`,
 		},
 	}, next)
@@ -107,9 +108,9 @@ func TestWebsocketSubscriptionClientImmediateClientCancel(t *testing.T) {
 		WithLogger(logger()),
 	)
 	next := make(chan []byte)
-	err := client.Subscribe(ctx, GraphQLSubscriptionOptions{
+	err := client.Subscribe(ctx, graphql_datasource.GraphQLSubscriptionOptions{
 		URL: strings.Replace(server.URL, "http", "ws", -1),
-		Body: GraphQLBody{
+		Body: graphql_datasource.GraphQLBody{
 			Query: `subscription {messageAdded(roomName: "room"){text}}`,
 		},
 	}, next)
@@ -161,9 +162,9 @@ func TestWebsocketSubscriptionClientWithServerDisconnect(t *testing.T) {
 		WithLogger(logger()),
 	)
 	next := make(chan []byte)
-	err := client.Subscribe(ctx, GraphQLSubscriptionOptions{
+	err := client.Subscribe(ctx, graphql_datasource.GraphQLSubscriptionOptions{
 		URL: strings.Replace(server.URL, "http", "ws", -1),
-		Body: GraphQLBody{
+		Body: graphql_datasource.GraphQLBody{
 			Query: `subscription {messageAdded(roomName: "room"){text}}`,
 		},
 	}, next)
@@ -218,9 +219,9 @@ func TestWebsocketSubscriptionClientErrorArray(t *testing.T) {
 		WithLogger(logger()),
 	)
 	next := make(chan []byte)
-	err := client.Subscribe(clientCtx, GraphQLSubscriptionOptions{
+	err := client.Subscribe(clientCtx, graphql_datasource.GraphQLSubscriptionOptions{
 		URL: strings.Replace(server.URL, "http", "ws", -1),
-		Body: GraphQLBody{
+		Body: graphql_datasource.GraphQLBody{
 			Query: `subscription {messageAdded(roomNam: "room"){text}}`,
 		},
 	}, next)
@@ -270,9 +271,9 @@ func TestWebsocketSubscriptionClientErrorObject(t *testing.T) {
 		WithLogger(logger()),
 	)
 	next := make(chan []byte)
-	err := client.Subscribe(clientCtx, GraphQLSubscriptionOptions{
+	err := client.Subscribe(clientCtx, graphql_datasource.GraphQLSubscriptionOptions{
 		URL: strings.Replace(server.URL, "http", "ws", -1),
-		Body: GraphQLBody{
+		Body: graphql_datasource.GraphQLBody{
 			Query: `subscription {messageAdded(roomNam: "room"){text}}`,
 		},
 	}, next)
@@ -383,9 +384,9 @@ func TestWebsocketSubscriptionClientDeDuplication(t *testing.T) {
 
 	next := make(chan []byte)
 	ctx, clientCancel := context.WithCancel(context.Background())
-	err := client.Subscribe(ctx, GraphQLSubscriptionOptions{
+	err := client.Subscribe(ctx, graphql_datasource.GraphQLSubscriptionOptions{
 		URL: strings.Replace(server.URL, "http", "ws", -1),
-		Body: GraphQLBody{
+		Body: graphql_datasource.GraphQLBody{
 			Query: `subscription {messageAdded(roomName: "room"){text}}`,
 		},
 	}, next)
@@ -398,9 +399,9 @@ func TestWebsocketSubscriptionClientDeDuplication(t *testing.T) {
 
 		ctx, cancel := context.WithCancel(context.Background())
 
-		err := client.Subscribe(ctx, GraphQLSubscriptionOptions{
+		err := client.Subscribe(ctx, graphql_datasource.GraphQLSubscriptionOptions{
 			URL: strings.Replace(server.URL, "http", "ws", -1),
-			Body: GraphQLBody{
+			Body: graphql_datasource.GraphQLBody{
 				Query: `subscription {messageAdded(roomName: "room"){text}}`,
 			},
 		}, next)
